+++ CAInvoice/src/main/java/ua/its/slot7/caccounting/system/BEntityPresenter.java
@@ -50,7 +50,7 @@ public class BEntityPresenter implements BEntityPresenterAvatar {
 		mbST.add("invoice_prepared_for_phone", invoice.getPerson().getPhone());
 
 //		total
-		mbST.add("invoice_total", new Float(invoice.getSum()).toString());
+		mbST.add("invoice_total", new Float(invoice.getSum().floatValue()).toString());
 
 //		date of issue
 		mbST.add("invoice_date_of_issue", invoice.getDateCreation().toString());
